refactor (ontologies): rename property class file
