+++ src/models/v2/ontologies/property-class.ts
